Update dependencies, enhance agent functionality, and refactor example commands
This commit introduces the following changes:
- Added new dependencies: `github.com/cohesion-org/deepseek-go` and `github.com/theapemachine/qpool` to `go.mod` and `go.sum`.
- Implemented a new method `GetRole` in the `Agent` struct to retrieve the agent's role.
- Enhanced the `Compile` method in the `Context` struct to preserve existing messages and add a system message.
- Refactored example commands to utilize the new `qpool` for managing concurrent tasks, improving the structure and readability of the development and research pipelines.
- Removed the `system` package, including `Graph` and `Edge` structures, to streamline the codebase and reduce complexity.

These changes improve the overall functionality and maintainability of the code, while also enhancing the user experience in example scenarios.

+++ provider/deepseek.go
@@ -49,10 +49,10 @@ func (d *DeepSeek) Generate(ctx context.Context, params *GenerationParams) <-cha
 
 		// Add optional parameters
 		if params.MaxTokens > 0 {
-			request.MaxTokens = params.MaxTokens
+			request.MaxTokens = int(params.MaxTokens)
 		}
 		if params.Temperature > 0 {
-			request.Temperature = params.Temperature
+			request.Temperature = float32(params.Temperature)
 		}
 
 		stream, err := d.client.CreateChatCompletionStream(ctx, request)
